Sync with latest Solr

+++ solr/core/src/java/org/apache/solr/schema/ManagedIndexSchemaFactory.java
@@ -35,6 +35,7 @@ import org.apache.solr.core.SolrCore;
 import org.apache.solr.core.SolrResourceLoader;
 import org.apache.solr.util.SystemIdResolver;
 import org.apache.solr.util.plugin.SolrCoreAware;
+import org.apache.zookeeper.CreateMode;
 import org.apache.zookeeper.KeeperException;
 import org.apache.zookeeper.data.Stat;
 import org.slf4j.Logger;
@@ -44,7 +45,7 @@ import org.xml.sax.InputSource;
 /** Factory for ManagedIndexSchema */
 public class ManagedIndexSchemaFactory extends IndexSchemaFactory implements SolrCoreAware {
   private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
-  private static final String UPGRADED_SCHEMA_EXTENSION = ".bak";
+  public static final String UPGRADED_SCHEMA_EXTENSION = ".bak";
   private static final String SCHEMA_DOT_XML = "schema.xml";
 
   public static final String DEFAULT_MANAGED_SCHEMA_RESOURCE_NAME = "managed-schema";
@@ -67,7 +68,7 @@ public class ManagedIndexSchemaFactory extends IndexSchemaFactory implements Sol
 
   @Override
   public void init(NamedList args) {
-    SolrParams params = SolrParams.toSolrParams(args);
+    SolrParams params = args.toSolrParams();
     isMutable = params.getBool("mutable", true);
     args.remove("mutable");
     managedSchemaResourceName = params.get(MANAGED_SCHEMA_RESOURCE_NAME, DEFAULT_MANAGED_SCHEMA_RESOURCE_NAME);
@@ -83,6 +84,11 @@ public class ManagedIndexSchemaFactory extends IndexSchemaFactory implements Sol
       log.error(msg);
       throw new SolrException(ErrorCode.SERVER_ERROR, msg);
     }
+  }
+
+  @Override
+  public String getSchemaResourceName(String cdResourceName) {
+    return managedSchemaResourceName; // actually a guess; reality depends on the actual files in the config set :-(
   }
 
   /**
@@ -143,7 +149,7 @@ public class ManagedIndexSchemaFactory extends IndexSchemaFactory implements Sol
       if (null == schemaInputStream) {
         // The managed schema file could not be found - load the non-managed schema
         try {
-          schemaInputStream = loader.openSchema(resourceName);
+          schemaInputStream = loader.openResource(resourceName);
           loadedResource = resourceName;
           shouldUpgrade = true;
         } catch (Exception e) {
@@ -168,22 +174,13 @@ public class ManagedIndexSchemaFactory extends IndexSchemaFactory implements Sol
     }
     InputSource inputSource = new InputSource(schemaInputStream);
     inputSource.setSystemId(SystemIdResolver.createSystemIdFromResourceName(loadedResource));
-    try {
-      schema = new ManagedIndexSchema(config, loadedResource, inputSource, isMutable, 
-                                      managedSchemaResourceName, schemaZkVersion, getSchemaUpdateLock());
-    } catch (KeeperException e) {
-      final String msg = "Error instantiating ManagedIndexSchema";
-      log.error(msg, e);
-      throw new SolrException(SolrException.ErrorCode.SERVER_ERROR, msg, e);
-    } catch (InterruptedException e) {
-      // Restore the interrupted status
-      Thread.currentThread().interrupt();
-      log.warn("", e);
-    }
-
+    schema = new ManagedIndexSchema(config, loadedResource, inputSource, isMutable,
+                                    managedSchemaResourceName, schemaZkVersion, getSchemaUpdateLock());
     if (shouldUpgrade) {
       // Persist the managed schema if it doesn't already exist
-      upgradeToManagedSchema();
+      synchronized (schema.getSchemaUpdateLock()) {
+        upgradeToManagedSchema();
+      }
     }
 
     return schema;
@@ -193,7 +190,7 @@ public class ManagedIndexSchemaFactory extends IndexSchemaFactory implements Sol
     InputStream schemaInputStream = null;
     try {
       // Attempt to load the managed schema
-      schemaInputStream = loader.openSchema(managedSchemaResourceName);
+      schemaInputStream = loader.openResource(managedSchemaResourceName);
       loadedResource = managedSchemaResourceName;
       warnIfNonManagedSchemaExists();
     } catch (IOException e) {
@@ -203,7 +200,7 @@ public class ManagedIndexSchemaFactory extends IndexSchemaFactory implements Sol
     if (null == schemaInputStream) {
       // The managed schema file could not be found - load the non-managed schema
       try {
-        schemaInputStream = loader.openSchema(resourceName);
+        schemaInputStream = loader.openResource(resourceName);
         loadedResource = resourceName;
         shouldUpgrade = true;
       } catch (Exception e) {
@@ -238,7 +235,7 @@ public class ManagedIndexSchemaFactory extends IndexSchemaFactory implements Sol
       } else { // Config is not in ZooKeeper
         InputStream nonManagedSchemaInputStream = null;
         try {
-          nonManagedSchemaInputStream = loader.openSchema(resourceName);
+          nonManagedSchemaInputStream = loader.openResource(resourceName);
           if (null != nonManagedSchemaInputStream) {
             exists = true;
           }
@@ -324,20 +321,32 @@ public class ManagedIndexSchemaFactory extends IndexSchemaFactory implements Sol
    * and no exception will be thrown.
    */
   private void zkUgradeToManagedSchema() {
-    schema.persistManagedSchemaToZooKeeper(true); // Only create, don't update it if it already exists
-
-    // After successfully persisting the managed schema, rename the non-managed
-    // schema znode by appending UPGRADED_SCHEMA_EXTENSION to its name.
-
     if (resourceName.equals(managedSchemaResourceName)) {
       log.info("On upgrading to managed schema, did not rename non-managed schema "
           + resourceName + " because it's the same as the managed schema's name.");
-    } else {
+      return;
+    }
+    final ZkSolrResourceLoader zkLoader = (ZkSolrResourceLoader)loader;
+    final ZkController zkController = zkLoader.getZkController();
+    final SolrZkClient zkClient = zkController.getZkClient();
+    final String lockPath = zkLoader.getConfigSetZkPath() + "/schemaUpgrade.lock";
+    boolean locked = false;
+    try {
+      try {
+        zkClient.makePath(lockPath, null, CreateMode.EPHEMERAL, null, true, true);
+        locked = true;
+      } catch (Exception e) {
+        // some other node already started the upgrade, or an error occurred - bail out
+        return;
+      }
+      schema.persistManagedSchemaToZooKeeper(true); // Only create, don't update it if it already exists
+
+      // After successfully persisting the managed schema, rename the non-managed
+      // schema znode by appending UPGRADED_SCHEMA_EXTENSION to its name.
+
       // Rename the non-managed schema znode in ZooKeeper
-      ZkSolrResourceLoader zkLoader = (ZkSolrResourceLoader)loader;
       final String nonManagedSchemaPath = zkLoader.getConfigSetZkPath() + "/" + resourceName;
       try {
-        ZkController zkController = zkLoader.getZkController();
         ZkCmdExecutor zkCmdExecutor = new ZkCmdExecutor(zkController.getClientTimeout());
         if (zkController.pathExists(nonManagedSchemaPath)) {
           // First, copy the non-managed schema znode content to the upgraded schema znode
@@ -346,16 +355,22 @@ public class ManagedIndexSchemaFactory extends IndexSchemaFactory implements Sol
           zkCmdExecutor.ensureExists(upgradedSchemaPath, zkController.getZkClient());
           zkController.getZkClient().setData(upgradedSchemaPath, bytes, true);
           // Then delete the non-managed schema znode
-          zkController.getZkClient().delete(nonManagedSchemaPath, -1, true);
+          if (zkController.getZkClient().exists(nonManagedSchemaPath, true)) {
+            try {
+              zkController.getZkClient().delete(nonManagedSchemaPath, -1, true);
+            } catch (KeeperException.NoNodeException ex) {
+              // ignore - someone beat us to it
+            }
+          }
 
-          // Set the resource name to the managed schema so that the CoreAdminHandler returns a findable filename 
+          // Set the resource name to the managed schema so that the CoreAdminHandler returns a findable filename
           schema.setResourceName(managedSchemaResourceName);
 
           log.info("After upgrading to managed schema in ZooKeeper, renamed the non-managed schema "
-                  + nonManagedSchemaPath + " to " + upgradedSchemaPath);
+              + nonManagedSchemaPath + " to " + upgradedSchemaPath);
         } else {
           log.info("After upgrading to managed schema in ZooKeeper, the non-managed schema "
-                  + nonManagedSchemaPath + " no longer exists.");
+              + nonManagedSchemaPath + " no longer exists.");
         }
       } catch (Exception e) {
         if (e instanceof InterruptedException) {
@@ -363,6 +378,17 @@ public class ManagedIndexSchemaFactory extends IndexSchemaFactory implements Sol
         }
         final String msg = "Error persisting managed schema resource " + managedSchemaResourceName;
         log.warn(msg, e); // Log as warning and suppress the exception
+      }
+    } finally {
+      if (locked) {
+        // unlock
+        try {
+          zkClient.delete(lockPath, -1, true);
+        } catch (KeeperException.NoNodeException nne) {
+          // ignore - someone else deleted it
+        } catch (Exception e) {
+          log.warn("Unable to delete schema upgrade lock file " + lockPath, e);
+        }
       }
     }
   }
