+++ solr/core/src/test/org/apache/solr/rest/schema/analysis/TestManagedSynonymFilterFactory.java
@@ -15,7 +15,9 @@
  * limitations under the License.
  */
 package org.apache.solr.rest.schema.analysis;
+
 import java.io.File;
+import java.net.URLEncoder;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
@@ -30,8 +32,9 @@ import org.eclipse.jetty.servlet.ServletHolder;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.noggit.JSONUtil;
 import org.restlet.ext.servlet.ServerServlet;
+
+import static org.apache.solr.common.util.Utils.toJSONString;
 
 public class TestManagedSynonymFilterFactory extends RestTestBase {
   
@@ -58,9 +61,14 @@ public class TestManagedSynonymFilterFactory extends RestTestBase {
 
   @After
   private void after() throws Exception {
-    jetty.stop();
-    jetty = null;
-    FileUtils.deleteDirectory(tmpSolrHome);
+    if (null != jetty) {
+      jetty.stop();
+      jetty = null;
+    }
+    if (null != tmpSolrHome) {
+      FileUtils.deleteDirectory(tmpSolrHome);
+      tmpSolrHome = null;
+    }
     System.clearProperty("managed.schema.mutable");
     System.clearProperty("enable.update.log");
     
@@ -86,8 +94,8 @@ public class TestManagedSynonymFilterFactory extends RestTestBase {
     // put a new mapping into the synonyms
     Map<String,List<String>> syns = new HashMap<>();
     syns.put("happy", Arrays.asList("glad","cheerful","joyful"));    
-    assertJPut(endpoint, 
-               JSONUtil.toJSON(syns),
+    assertJPut(endpoint,
+        toJSONString(syns),
                "/responseHeader/status==0");
     
     assertJQ(endpoint,
@@ -112,8 +120,8 @@ public class TestManagedSynonymFilterFactory extends RestTestBase {
     syns = new HashMap<>();
     syns.put("sad", Arrays.asList("unhappy"));    
     syns.put("SAD", Arrays.asList("bummed"));    
-    assertJPut(endpoint, 
-               JSONUtil.toJSON(syns),
+    assertJPut(endpoint,
+        toJSONString(syns),
                "/responseHeader/status==0");
     
     assertJQ(endpoint, 
@@ -154,23 +162,50 @@ public class TestManagedSynonymFilterFactory extends RestTestBase {
             "count(/response/lst[@name='field']) = 1",
             "/response/lst[@name='responseHeader']/int[@name='status'] = '0'");
 
+    // multi-term synonym logic - SOLR-10264
+    final String multiTermOrigin;
+    final String multiTermSynonym;
+    if (random().nextBoolean()) {
+      multiTermOrigin  = "hansestadt hamburg";
+      multiTermSynonym = "hh";
+    } else {
+      multiTermOrigin  = "hh";
+      multiTermSynonym = "hansestadt hamburg";
+    }
+    // multi-term logic similar to the angry/mad logic (angry ~ origin, mad ~ synonym)
+
     assertU(adoc(newFieldName, "I am a happy test today but yesterday I was angry", "id", "5150"));
+    assertU(adoc(newFieldName, multiTermOrigin+" is in North Germany.", "id", "040"));
     assertU(commit());
 
     assertQ("/select?q=" + newFieldName + ":angry",
             "/response/lst[@name='responseHeader']/int[@name='status'] = '0'",
             "/response/result[@name='response'][@numFound='1']",
             "/response/result[@name='response']/doc/str[@name='id'][.='5150']");    
+    assertQ("/select?q=" + newFieldName + ":"+URLEncoder.encode(multiTermOrigin, "UTF-8"),
+        "/response/lst[@name='responseHeader']/int[@name='status'] = '0'",
+        "/response/result[@name='response'][@numFound='1']",
+        "/response/result[@name='response']/doc/str[@name='id'][.='040']");
     
     // add a mapping that will expand a query for "mad" to match docs with "angry"
     syns = new HashMap<>();
     syns.put("mad", Arrays.asList("angry"));    
-    assertJPut(endpoint, 
-               JSONUtil.toJSON(syns),
+    assertJPut(endpoint,
+        toJSONString(syns),
                "/responseHeader/status==0");
     
     assertJQ(endpoint, 
         "/synonymMappings/managedMap/mad==['angry']");
+
+    // add a mapping that will expand a query for "multi-term synonym" to match docs with "acronym"
+    syns = new HashMap<>();
+    syns.put(multiTermSynonym, Arrays.asList(multiTermOrigin));
+    assertJPut(endpoint,
+        toJSONString(syns),
+               "/responseHeader/status==0");
+
+    assertJQ(endpoint+"/"+URLEncoder.encode(multiTermSynonym, "UTF-8"),
+        "/"+multiTermSynonym+"==['"+multiTermOrigin+"']");
 
     // should not match as the synonym mapping between mad and angry does not    
     // get applied until core reload
@@ -178,6 +213,12 @@ public class TestManagedSynonymFilterFactory extends RestTestBase {
         "/response/lst[@name='responseHeader']/int[@name='status'] = '0'",
         "/response/result[@name='response'][@numFound='0']");    
     
+    // should not match as the synonym mapping between "origin" and "synonym"
+    // was not added before the document was indexed
+    assertQ("/select?q=" + newFieldName + ":("+URLEncoder.encode(multiTermSynonym, "UTF-8") + ")&sow=false",
+        "/response/lst[@name='responseHeader']/int[@name='status'] = '0'",
+        "/response/result[@name='response'][@numFound='0']");
+
     restTestHarness.reload();
 
     // now query for mad and we should see our test doc
@@ -186,16 +227,22 @@ public class TestManagedSynonymFilterFactory extends RestTestBase {
         "/response/result[@name='response'][@numFound='1']",
         "/response/result[@name='response']/doc/str[@name='id'][.='5150']");    
     
+    // now query for "synonym" and we should see our test doc with "origin"
+    assertQ("/select?q=" + newFieldName + ":("+URLEncoder.encode(multiTermSynonym, "UTF-8") + ")&sow=false",
+        "/response/lst[@name='responseHeader']/int[@name='status'] = '0'",
+        "/response/result[@name='response'][@numFound='1']",
+        "/response/result[@name='response']/doc/str[@name='id'][.='040']");
+
     // test for SOLR-6015
     syns = new HashMap<>();
     syns.put("mb", Arrays.asList("megabyte"));    
     assertJPut(endpoint,
-        JSONUtil.toJSON(syns),
+        toJSONString(syns),
         "/responseHeader/status==0");
 
     syns.put("MB", Arrays.asList("MiB", "Megabyte"));    
     assertJPut(endpoint,
-        JSONUtil.toJSON(syns),
+        toJSONString(syns),
         "/responseHeader/status==0");
     
     assertJQ(endpoint + "/MB",
@@ -204,7 +251,7 @@ public class TestManagedSynonymFilterFactory extends RestTestBase {
     // test for SOLR-6878 - by default, expand is true, but only applies when sending in a list
     List<String> m2mSyns = new ArrayList<>();
     m2mSyns.addAll(Arrays.asList("funny", "entertaining", "whimiscal", "jocular"));
-    assertJPut(endpoint, JSONUtil.toJSON(m2mSyns), "/responseHeader/status==0");
+    assertJPut(endpoint, toJSONString(m2mSyns), "/responseHeader/status==0");
 
     assertJQ(endpoint + "/funny",
         "/funny==['entertaining','funny','jocular','whimiscal']");
@@ -236,7 +283,7 @@ public class TestManagedSynonymFilterFactory extends RestTestBase {
     // now put a synonym
     syns.put("fröhlich", Arrays.asList("glücklick"));
     assertJPut(endpoint,
-        JSONUtil.toJSON(syns),
+        toJSONString(syns),
         "/responseHeader/status==0");
 
     // and check if it exists
